requirements.txt
--- requirements.txt ---
fastapi
uvicorn[standard]
numpy
pydub
soundfile
librosa
python-multipart
websockets
openai-whisper